requirements.txt
python-dotenv
requests
pandas
numpy
matplotlib
tensorflow==2.13.0
scikit-learn